191007|django_double_app

+++ 03_Django/04_django_review/articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from django.core.exceptions import ValidationError
 from .models import Article
+from IPython import embed
 
 # Create your views here.
 def index(request):
@@ -11,20 +12,20 @@ def index(request):
     context = {'articles': articles}
     return render(request, 'articles/index.html', context)
 
-def new(request):
-    return render(request,'articles/new.html')
+# def new(request):
+#     # embed()
+#     return render(request, 'articles/new.html')
 
 def create(request):
-    try:
+    if request.method == 'POST':
         title = request.POST.get('title')
         content = request.POST.get('content')
-        article = Article(title=title, content = content)
+        article = Article(title=title, content=content)
         article.full_clean()
-    except ValidationError:
-        raise ValidationError('Your Error message')
-    else:
         article.save()
         return redirect('articles:detail', article.pk)
+    else:
+        return render(request, 'articles/create.html')
     """
     title = request.POST.get('title')
     # print(title)
@@ -56,18 +57,24 @@ def detail(request, pk):    #pk: primary key
   
 def delete(request,pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
-    
-    return redirect('articles:index')
+    if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return render(request, 'articles/index.html')
 
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {'article': article}
-    return render(request, 'articles/edit.html', context)
+# def edit(request, pk):
+#     article = Article.objects.get(pk=pk)
+#     context = {'article': article}
+#     return render(request, 'articles/edit.html', context)
 
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    article.title = request.POST.get('title')
-    article.content = request.POST.get('content')
-    article.save()
-    return redirect('articles:detail', article.pk)
+    if request.method == 'POST':
+        article.title = request.POST.get('title')
+        article.content = request.POST.get('content')
+        article.save()
+        return redirect('articles:detail', article.pk)
+    else:
+        context = {'article': article}
+        return render(request, 'articles/update.html', context)
